Thread tutorial reorganization
- this groups both thread tutorials to reflect the tock book structure.
