--- examples/tutorials/thread_network/07_screen_button/Makefile
@@ -1,13 +1,0 @@
-# Makefile for user application
-
-# Specify this directory relative to the current application.
-TOCK_USERLAND_BASE_DIR = ../../../../
-
-# Which files to compile.
-C_SRCS := $(wildcard *.c)
-
-PACKAGE_NAME = screen
-
-# Include userland master makefile. Contains rules and flags for actually
-# building the application.
-include $(TOCK_USERLAND_BASE_DIR)/AppMakefile.mk
