+++ examples/tutorials/thread_tutorials/encrypted_sensor_data/00_thread_network/README.md
