+++ examples/tutorials/thread_tutorials/encrypted_sensor_data/01_openthread_attach/README.md
