+++ examples/tutorials/thread_tutorials/encrypted_sensor_data/02_openthread_recv/main.c
