+++ examples/tutorials/thread_tutorials/encrypted_sensor_data/03_encrypted_data/Makefile
@@ -1,0 +1,19 @@
+# Makefile for user application
+
+# Specify this directory relative to the current application.
+TOCK_USERLAND_BASE_DIR = ../../../../..
+
+# Specify this app depends on the MTD OpenThread library.
+include $(TOCK_USERLAND_BASE_DIR)/libopenthread/libopenthread-mtd.mk
+
+# set stack size to 8000 to support openthread app
+STACK_SIZE:=8000
+
+# App name
+PACKAGE_NAME:=openthread_sensor
+
+C_SRCS := main.c oracle.c
+
+# Include userland master makefile. Contains rules and flags for actually
+# building the application.
+include $(TOCK_USERLAND_BASE_DIR)/AppMakefile.mk
