+++ examples/tutorials/thread_tutorials/encrypted_sensor_data/03_encrypted_data/main.c
