+++ examples/tutorials/thread_tutorials/encrypted_sensor_data/README.md
