+++ examples/tutorials/thread_tutorials/temperature_sensor/00_sensor_hello/main.c
