+++ examples/tutorials/thread_tutorials/temperature_sensor/01_sensor_temperature/main.c
