+++ examples/tutorials/thread_tutorials/temperature_sensor/02_sensor_final/main.c
