+++ examples/tutorials/thread_tutorials/temperature_sensor/03_openthread/main.c
