+++ examples/tutorials/thread_tutorials/temperature_sensor/04_openthread_attach/main.c
