+++ examples/tutorials/thread_tutorials/temperature_sensor/05_openthread_final/main.c
