+++ examples/tutorials/thread_tutorials/temperature_sensor/06_screen/Makefile
@@ -1,0 +1,13 @@
+# Makefile for user application
+
+# Specify this directory relative to the current application.
+TOCK_USERLAND_BASE_DIR = ../../../../..
+
+# Which files to compile.
+C_SRCS := $(wildcard *.c)
+
+PACKAGE_NAME = screen
+
+# Include userland master makefile. Contains rules and flags for actually
+# building the application.
+include $(TOCK_USERLAND_BASE_DIR)/AppMakefile.mk
