+++ examples/tutorials/thread_tutorials/temperature_sensor/06_screen/main.c
