+++ examples/tutorials/thread_tutorials/temperature_sensor/07_screen_button/main.c
