+++ examples/tutorials/thread_tutorials/temperature_sensor/08_screen_u8g2/main.c
