+++ examples/tutorials/thread_tutorials/temperature_sensor/09_screen_final/main.c
