+++ examples/tutorials/thread_tutorials/temperature_sensor/10_screen_ipc/main.c
