+++ examples/tutorials/thread_tutorials/temperature_sensor/11_sensor_ipc/main.c
