+++ examples/tutorials/thread_tutorials/temperature_sensor/12_openthread_ipc/main.c
